Cleanup of build-image scripts
This PR re-arranges tags int GitHub actions to be more explicit in
the pull/build/push steps on which tag of the image is currently
pulled/pushed/built.

Previously the actual configuration of those steps was determined
by combining "top", "job" and "step" variables, but that gave
little insight on what combinations of evn variables was used in
the step eventually. After this change, in all the pull/push/build
steps you can see what environment is used for that build.

Also, since we often build and push and then pull the same image in
the same workflow, the PUSH tag was sometimes used in PULL
variable as input. We change it now by introducing single
`IMAGE_TAG_FOR_THE_BUILD` which will be used as PUSH and PULL
tag as needed.

This one also fixes failure of PROD cache builds which did not
pull the right CI image.